Split Angelscript engine into separate files
