+++ wxChaos/scripting/AngelscriptConfigurationEngine.h
@@ -1,0 +1,26 @@
+#pragma once
+
+#include <angelscript.h>
+#include <wx/string.h>
+#include "ScriptData.h"
+#include "EngineStatus.h"
+
+class AngelscriptConfigurationEngine
+{
+private:
+    ScriptData configuration;
+    asIScriptEngine* engine;
+    asIScriptContext* ctx;
+    EngineStatus status;
+    wxString errorInfo;
+    std::string filePath;
+public:
+    AngelscriptConfigurationEngine();
+    ~AngelscriptConfigurationEngine();
+
+    bool CompileFromPath(std::string path);
+    bool Execute();
+    ScriptData GetScriptData();
+    EngineStatus GetStatus();
+    wxString GetErrorInfo();
+};
