+++ wxChaos/scripting/EngineStatus.h
@@ -1,0 +1,7 @@
+#pragma once
+
+enum class EngineStatus
+{
+    Ok,
+    Error
+};
