+++ wxChaos/scripting/ScriptCategory.h
@@ -1,0 +1,10 @@
+#pragma once
+
+enum class ScriptCategory
+{
+    Complex,
+    NumMet,
+    Physic,
+    Other,
+    Undefined
+};
