+++ wxChaos/scripting/ScriptDataType.h
@@ -1,0 +1,7 @@
+#pragma once
+
+enum class ScriptDataType
+{
+    Standard,
+    Error
+};
